adk-multiagent-test/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adk-multiagent-test"
version = "0.1.0"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
  "google-adk>=1.0.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "jsonschema>=4.21",
  "opentelemetry-api>=1.22",
  "opentelemetry-sdk>=1.22",
  "opentelemetry-exporter-gcp-trace>=1.6",
  "opentelemetry-propagator-gcp>=1.6",
  "opentelemetry-instrumentation-httpx>=0.43b0",
  "opentelemetry-semantic-conventions>=0.43b0",
  "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
